fix caching purging

+++ bucketlist/src/app/app.component.ts
@@ -52,6 +52,7 @@ export class AppComponent implements OnInit  {
     public commonId:number = 0;
     public reportData:any;
     public usersData:any;
+    public fresh:any;
     public addData:any;
     public doneData:any;
     public writeTimeout:any;
@@ -108,6 +109,7 @@ export class AppComponent implements OnInit  {
                             this.selectLang((user && user.language)?user.language:'en');
                             this._cacheService.set('user_', user, {maxAge: 3 * 24 * 60 * 60});
                             this.broadcaster.broadcast('getUser', user);
+                            this.purgeFresh();
                         },
                         error => localStorage.removeItem('apiKey'));
             } else {
@@ -115,6 +117,7 @@ export class AppComponent implements OnInit  {
             }
         }
 
+        this.purgeFresh();
         if(this.appUser) {
             this.updatedEmail = this._cacheService.get('confirmRegEmail' + this.appUser.id);
 
@@ -131,6 +134,7 @@ export class AppComponent implements OnInit  {
         this.broadcaster.on<User>('login')
             .subscribe(user => {
               this.appUser = user;
+              this.purgeFresh();
               this.selectLang((this.appUser.language)?this.appUser.language:'en');
               this._cacheService.set('user_', user, {maxAge: 3 * 24 * 60 * 60});
               this.broadcaster.broadcast('getUser', user);
@@ -231,7 +235,7 @@ export class AppComponent implements OnInit  {
                       this.busy = false;
                       if(result){
                           if(result.remove){
-                              if(data.userGoal.goal.author == this.appUser.id){
+                              if(data.userGoal.goal.author && data.userGoal.goal.author.id == this.appUser.id){
                                   this.angulartics2.eventTrack.next({ action: 'Goal delete', properties: { category: 'Goal', label: 'Goal delete from angular2'}});
                               }else {
                                   this.angulartics2.eventTrack.next({ action: 'Goal unlisted', properties: { category: 'Goal', label: 'Goal unlisted from angular2'}});
@@ -249,6 +253,8 @@ export class AppComponent implements OnInit  {
                           this.broadcaster.broadcast('addGoal', data.userGoal);
                           this.broadcaster.broadcast('addGoal'+data.userGoal.goal.id, data.userGoal);
                       }
+
+                      this.testCache(data.userGoal.goal.id);
                   });
                   // this.addModal = true;
               });
@@ -281,10 +287,54 @@ export class AppComponent implements OnInit  {
                   dialogRef.componentInstance.userGoal = data.userGoal;
                   dialogRef.afterClosed().subscribe(result => {
                       this.broadcaster.broadcast('doneGoal'+data.userGoal.goal.id, {});
+                      this.testCache(data.userGoal.goal.id);
                   });
                   // this.doneModal = true;
               });
 
+    }
+
+    purgeFresh(){
+        if(this.appUser) {
+            this.fresh = this._cacheService.get('fresh'+this.appUser.id);
+            if(!this.fresh){
+                this.fresh = {
+                    'activities':true,
+                    'featuredIdea':true,
+                    'topIdea':true
+                }
+            }
+        }
+    }
+
+    testCache(id){
+        let featuredIdea = this._cacheService.get('featuredIdea');
+        let topIdea = this._cacheService.get('topIdea');
+        let activities = this._cacheService.get('activities' + this.appUser.id);
+        this.changeCache(featuredIdea, 'featuredIdea', id);
+        this.changeCache(topIdea, 'topIdea', id);
+        this.changeCache(activities, 'activities', id);
+        this._cacheService.set('fresh'+this.appUser.id, this.fresh);
+    }
+
+    changeCache(dates, name, id){
+        if(dates && dates.length){
+            for(let data of dates){
+                if(data.goals){
+                    for(let goal of data.goals){
+                        if(goal.id == id){
+                            this.fresh[name] = false;
+                            return;
+                        }
+                    }
+                } else {
+                    if(data.id == id){
+                        this.fresh[name] = false;
+                        return;
+                    }
+                }
+            }
+        }
     }
 
     toogleNote(){
